최종

+++ src/main/java/community/controller/DeleteCommunityServlet.java
@@ -12,7 +12,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import community.model.service.BoardService;
+import community.model.vo.CMBaordHashtag;
 import community.model.vo.CMBoardPhoto;
+import community.model.vo.Comment;
 
 /**
  * Servlet implementation class DeleteBoardServlet
@@ -36,18 +38,61 @@ public class DeleteCommunityServlet extends HttpServlet {
 			throws ServletException, IOException {
 		BoardService bservice = new BoardService();
 		RequestDispatcher view = null;
-		int boardNo = Integer.parseInt(request.getParameter("board"));
+		int boardNo = Integer.parseInt(request.getParameter("boardNo"));
 		
 		ArrayList<CMBoardPhoto> list = bservice.selectBoardPhotoList(boardNo);
 		
 		for(CMBoardPhoto photo : list) {
-			String fileName = photo.getFilename();
-			if(fileName != null) {
-				String savePath = request.getSession().getServletContext()
-						.getRealPath("/resources/board/images");
-				new File(savePath + "\\" + fileName).delete();
+			int result = bservice.deleteBoardPhoto(photo);
+			
+			if(result > 0) {
+				String fileName = photo.getFilename();
+				if(fileName != null) {
+					String savePath = request.getSession().getServletContext()
+							.getRealPath("/resources/board/images");
+					new File(savePath + "\\" + fileName).delete();
+					System.out.println(fileName + " 파일 삭제 성공");
+				}
+			} else {
+				view = request.getRequestDispatcher("views/common/error.jsp");
+				request.setAttribute("message", photo.getBoardNo() + "번 게시글 " + photo.getPhotoNo() + "번 사진 삭제 실패!");
+				view.forward(request, response);
 			}
 		}
+		
+		ArrayList<Comment> clist = bservice.selectCommentList(boardNo);
+		
+		if(clist.size() > 0) {
+			for(Comment comment : clist) {
+				int result = bservice.deleteComment(comment);
+				
+				if(result > 0) {
+					System.out.println(comment + " 삭제 성공");
+				} else {
+					view = request.getRequestDispatcher("views/common/error.jsp");
+					request.setAttribute("message", comment.getBoardNo() + "번 게시글 " + comment.getCommentNo() + "번 댓글 삭제 실패!");
+					view.forward(request, response);
+				}
+			}
+		}
+		
+		
+		ArrayList<CMBaordHashtag> bhlist = bservice.selectBoardHashtagList(boardNo);
+		
+		if(bhlist.size() > 0) {
+			for(CMBaordHashtag boardHashtag : bhlist) {
+				int result = bservice.deleteBoardHashtag(boardHashtag);
+				
+				if(result > 0) {
+					System.out.println(boardHashtag.getBoardNo() + "번 게시글의 " + boardHashtag.getHashtagNo() + "번 해시태그 삭제 성공");
+				} else {
+					view = request.getRequestDispatcher("views/common/error.jsp");
+					request.setAttribute("message", boardHashtag.getBoardNo() + "번 게시글의 " + boardHashtag.getHashtagNo() + "번 해시태그 삭제 실패!");
+					view.forward(request, response);
+				}
+			}
+		}
+		
 		
 		int result = bservice.deleteBoard(boardNo);
 		if(result > 0) {
